Move fake api into testing file for readability
