--- lambdas/find_new_granules/tests/fake_api.py
@@ -1,16 +1,0 @@
-
-
-def cmr_requests_get(*args, **kwargs):
-    """This method will be used by the mock to replace requests.get"""
-    class MockResponse:
-        def __init__(self, json_data, status_code):
-            self.json_data = json_data
-            self.status_code = status_code
-
-        def json(self):
-            return self.json_data
-
-    if args[0] == 'http://someurl.com/test.json':
-        return MockResponse({"key1": "value1"}, 200)
-    elif args[0] == 'http://someotherurl.com/anothertest.json':
-        return MockResponse({"key2": "value2"}, 200)
