+++ lambdas/find_new_granules/tests/test_find_new_granules.py
@@ -1,8 +1,21 @@
 import unittest
 from unittest import mock
 
-from .fake_api import cmr_requests_get
 from find_new_granules.find_new_granules import get_new
+
+
+def cmr_requests_get(*args, **kwargs):
+    """This method will be used by the mock to replace requests.get"""
+    class MockResponse:
+        def __init__(self, json_data, status_code):
+            self.json_data = json_data
+            self.status_code = status_code
+
+        def json(self):
+            return self.json_data
+
+    if args[0] == 'http://someurl.com/test.json':
+        return MockResponse({"key1": "value1"}, 200)
 
 
 class TestFindNewGranules(unittest.TestCase):
